Fix rendering thumbnail when source footage is a sequence.
